% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-tidy-select.R
\name{args_tidy_select}
\alias{args_tidy_select}
\title{Argument type: tidy-select}
\description{
Note: This page is adapted from the \code{args_tidy_select} help topic from the
tidyselect package. See LICENSE-tidyselect.md for license information.

This page describes the \verb{<tidy-select>} argument modifier which indicates
the argument supports \strong{tidy selections}. Tidy selection provides a concise
dialect of R for selecting variables based on their names or properties.

Various qualifyr functions use tidy selection to select tables from data sets
and columns from tables.
}
\section{Overview of selection features}{


Tidyverse selections implement a dialect of R where operators make
it easy to select variables:
\itemize{
\item \code{:} for selecting a range of consecutive variables.
\item \code{!} for taking the complement of a set of variables.
\item \code{&} and \code{|} for selecting the intersection or the union of two sets of
variables.
\item \code{c()} for combining selections.
}

In addition, you can use \strong{selection helpers}. Some helpers select specific
columns:
\itemize{
\item \code{\link[tidyselect:everything]{everything()}}: Matches all variables.
\item \code{\link[tidyselect:everything]{last_col()}}: Select last variable, possibly with
an offset.
}

Other helpers select variables by matching patterns in their names:
\itemize{
\item \code{\link[tidyselect:starts_with]{starts_with()}}: Starts with a prefix.
\item \code{\link[tidyselect:starts_with]{ends_with()}}: Ends with a suffix.
\item \code{\link[tidyselect:starts_with]{contains()}}: Contains a literal string.
\item \code{\link[tidyselect:starts_with]{matches()}}: Matches a regular expression.
\item \code{\link[tidyselect:starts_with]{num_range()}}: Matches a numerical range like
x01, x02, x03.
}

Or from external variables stored in a character vector:
\itemize{
\item \code{\link[tidyselect:all_of]{all_of()}}: Matches variable names in a character
vector. All names must be present, otherwise an out-of-bounds error is
thrown.
\item \code{\link[tidyselect:all_of]{any_of()}}: Same as \code{all_of()}, except that no
error is thrown for names that don't exist.
}

Or using a predicate function:
\itemize{
\item \code{\link[tidyselect:where]{where()}}: Applies a function to all variables and
selects those for which the function returns \code{TRUE}.
}
}

\section{Extensions to tidy selection syntax in qualifyr}{


The \code{reference} argument of
\code{\link[=cstr_foreign_key]{cstr_foreign_key()}} supports two extensions
to tidyselect syntax:
\itemize{
\item The symbol \code{.self} refers to the same table
\item Specific columns from another table can be specified using the syntax
\code{table[columns]} or \code{table$column}
}
}

\seealso{
\code{vignette("syntax", package = "tidyselect")}
}
