% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_dataframe.R
\name{as_qf_dataframe}
\alias{as_qf_dataframe}
\title{Convert a data frame to a qualifyr data frame}
\usage{
as_qf_dataframe(x)
}
\arguments{
\item{x}{Any object inheriting from class \verb{<data.frame>}}
}
\value{
A copy of \code{x} with class \verb{<qf_dataframe>} (retaining previous class
information as superclasses) and the \code{constraints} attribute set to an
empty list.
}
\description{
Create a qualifyr data frame object from a data frame
}
