% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_table.R
\name{as_qf_table}
\alias{as_qf_table}
\title{Convert a data frame to a qualifyr table}
\usage{
as_qf_table(x)
}
\arguments{
\item{x}{Any object inheriting from class \verb{<data.frame>}}
}
\value{
A copy of \code{x} with class \verb{<qf_table>} (retaining previous class
information as superclasses) and the \code{constraints} attribute set to an
empty list.
}
\description{
Create a qualifyr table object from a data frame
}
