% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_constraint.R
\name{exceptions}
\alias{exceptions}
\alias{exceptions<-}
\title{Get or set the exceptions of a constraint}
\usage{
exceptions(x)

exceptions(x) <- value
}
\arguments{
\item{x}{A constraint object}

\item{value}{A list of exceptions}
}
\value{
For \code{exceptions}, the list of \verb{<qf_exception>} objects associated
with the constraint For \verb{exceptions<-}, an updated version of \code{x} with
exceptions set.
}
\description{
Get or set the exceptions of a constraint
}
