% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_constraint.R
\name{from}
\alias{from}
\alias{\%from\%}
\title{Specify a reference to columns from another table}
\usage{
cols \%from\% table
}
\arguments{
\item{cols}{<\code{tidy-select}> Reference columns}

\item{table}{<\code{tidy-select}> Reference table}

\item{.context}{A \verb{<qf_dataset}> object. Should not be supplied when \verb{\%from\%}
is used within a \verb{cstr_*} function argument.}
}
\value{
A character vector of column names, with the \code{table} attribute set.
}
\description{
The \verb{\%from\%} infix operator identifies columns from a specific table within
a qualifyr data set.
}
