% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_check.R
\name{inspect_results}
\alias{inspect_results}
\alias{satisfied}
\alias{handled}
\alias{as.logical.qf_check}
\title{Inspect constraint check results}
\usage{
satisfied(x)

handled(x)

\method{as.logical}{qf_check}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=check_constraints]{check_constraints()}}, i.e., a \verb{<qf_check}>
object}

\item{...}{Not used}
}
\value{
Logical; for \code{satisfied()}, whether all constraints were satisfied,
and for \code{handled()}, whether all constraints were either satisfied or
excepted. \code{as.logical(x)} is equivalent to \code{handled(x)}.
}
\description{
Functions to access attributes of constraint check results
}
