% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifyr-package.R
\docType{package}
\name{qualifyr-package}
\alias{qualifyr}
\alias{qualifyr-package}
\title{qualifyr: Check Constraints on Relational Data Structures}
\description{
Lorem ipsum dolor sit amet.
}
\author{
\strong{Maintainer}: Ian Farm \email{ian.farm@maine.edu}

}
\keyword{internal}
