% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_dataset.R
\name{subsetting}
\alias{subsetting}
\alias{$.qf_dataset}
\alias{[[.qf_dataset}
\alias{[.qf_dataset}
\alias{$<-.qf_dataset}
\alias{[[<-.qf_dataset}
\alias{[<-.qf_dataset}
\title{Subsetting qualifyr data sets}
\usage{
\method{$}{qf_dataset}(x, name)

\method{[[}{qf_dataset}(x, i)

\method{[}{qf_dataset}(x, i)

\method{$}{qf_dataset}(x, name) <- value

\method{[[}{qf_dataset}(x, i) <- value

\method{[}{qf_dataset}(x, i) <- value
}
\arguments{
\item{x}{A qualifyr data set}

\item{name}{The name of a table to subset, either as a character literal or
raw symbol}

\item{i}{Indices specifying tables - numeric for positional subsetting, or
character for subsetting by name; must be length one for \code{[[} and \verb{[[<-}}

\item{value}{A qualifyr table to replace the subsetted table(s) with;
or, for \verb{[<-}, possibly a list of qualifyr tables or a qualifyr dataset}
}
\value{
For \code{$} and \code{[[}, a qualifyr table. For \code{[}, a list of qualifyr
tables. For \verb{$<-}, \verb{[[<-}, and \verb{[<-}, a modified version of \code{x}.
}
\description{
Accessing tables from qualifyr data sets works like subsetting bare lists,
except that certain context information is attached to the returned tables.
This is necessary for resolving and checking foreign keys that reference
other tables in the dataset.
}
\details{
Subsetted tables have \code{x} as their \code{context} attribute. The subset assignment
methods strip the \code{context} attribute from \code{value}.
}
