% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_constraint.R
\name{apply_to_each}
\alias{apply_to_each}
\title{Apply the same constraint to multiple columns or sets of columns}
\usage{
apply_to_each(.cstr, ..., .args = list())
}
\arguments{
\item{.cstr}{A constraint specifier function (one of \code{\link[=cstr_]{cstr_*()}})}

\item{...}{<\code{\link[=args_tidy_select]{tidy-select}}> Each is passed to the \code{cols}
argument of \code{.cstr}}

\item{.args}{List of additional arguments passed to \code{.cstr}}
}
\value{
A \verb{<qf_constraint_list>}
}
\description{
\code{apply_to_each()} applies a constraint specifier function
(\code{\link[=cstr_]{cstr_*()}}) to each of several columns or sets of columns specified
in \code{...}
}
