% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/qf_dataframe.R, R/qf_dataset.R
\name{check_constraints}
\alias{check_constraints}
\alias{check_constraints.qf_dataframe}
\alias{check_constraints.qf_dataset}
\title{Check constraints on qualifyr data set or data frame}
\usage{
check_constraints(x, ...)

\method{check_constraints}{qf_dataframe}(x, dataset = NULL, ...)

\method{check_constraints}{qf_dataset}(x, ...)
}
\arguments{
\item{x}{A \verb{<qf_dataset>} or \verb{<qf_dataframe>} object}

\item{...}{Further arguments passed to methods}

\item{dataset}{If \code{x} is a data frame with foreign key constraints, the
data set to reference foreign keys against. Can be \code{NULL} if \code{x} has no
foreign keys.}
}
\description{
Checks that a data set or data frame satisfies the constraints set on it.
}
