% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/qf_dataset.R, R/qf_table.R
\name{check_constraints}
\alias{check_constraints}
\alias{check_constraints.qf_dataset}
\alias{check_constraints.qf_table}
\title{Check constraints on qualifyr data set or table}
\usage{
check_constraints(x, ...)

\method{check_constraints}{qf_dataset}(x, ...)

\method{check_constraints}{qf_table}(x, dataset = NULL, ...)
}
\arguments{
\item{x}{A \verb{<qf_dataset>} or \verb{<qf_table>} object}

\item{...}{Further arguments passed to methods}

\item{dataset}{If \code{x} is a table with foreign key constraints, the
data set to reference foreign keys against. Can be \code{NULL} if \code{x} has no
foreign keys.}
}
\description{
Checks that a data set or table satisfies the constraints set on it.
}
