% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/qf_dataset.R, R/qf_table.R
\name{check_constraints}
\alias{check_constraints}
\alias{check_constraints.qf_dataset}
\alias{check_constraints.qf_table}
\title{Check constraints on qualifyr data set or table}
\usage{
check_constraints(x)

\method{check_constraints}{qf_dataset}(x)

\method{check_constraints}{qf_table}(x)
}
\arguments{
\item{x}{A \verb{<qf_dataset>} or \verb{<qf_table>} object.}
}
\value{
If \code{x} is a \verb{<qf_table>} object, a list where each entry is the
results from checking one constraint. If \code{x} is a \verb{<qf_dataset>} object, a
list where each entry is the results from checking constraints on one
table. The result from checking one constraint takes the form of a list
with the elements:
\itemize{
\item \verb{$satsified} - A logical indicating whether the constraint was satisfied,
not considering exceptions
\item \verb{$handled} - A logical indicating whether every row either satisfied the
constraint or was excepted
\item \verb{$rows} - A list of logical vectors indicating whether each row:
\itemize{
\item \code{.. $satisfied} - satisfied the condition
\item \code{.. $excepted} - was excepted
\item \code{.. $handled} - satisfied the condition or was excepted
}
}
}
\description{
Checks that a data set or table satisfies the constraints set on it.
}
\details{
If \code{x} is a \verb{<qf_table>} with a foreign key that references another
table, then \code{x} must have the \code{context} attribute set. The \code{context}
attribute is set when \code{x} is accessed by subsetting a \verb{<qf_dataset>}
object.
}
