% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_table.R
\name{constraints}
\alias{constraints}
\alias{constraints<-}
\title{Get or set constraints of a qualifyr table}
\usage{
constraints(x)

constraints(x) <- value
}
\arguments{
\item{x}{A qualifyr table to set constraints on}

\item{value}{A list of constraints (or constraint specifiers), as returned
by \verb{cstr_*()} functions.}
}
\value{
For \code{constraints}, the list of \verb{<qf_constraint>} objects associated
with the table. For \verb{constraints<-}, an updated version of \code{x} with
constraints set.
}
\description{
Get or set constraints of a qualifyr table
}
\details{
If \code{value} contains foreign key specifiers referencing another
table, \code{x} must have the \code{context} attribute set, otherwise \verb{constraints<-}
will signal an error. Tables returned by subsetting a qualifyr dataset
using \code{$}, \code{[[}, or \code{[} have this attribute set.
}
\seealso{
\link{pick_constraint}
}
