% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/qf_dataset.R, R/qf_table.R
\name{constraints}
\alias{constraints}
\alias{constraints<-}
\alias{constraints.qf_dataset}
\alias{constraints<-.qf_dataset}
\alias{constraints.qf_table}
\alias{constraints<-.qf_table}
\title{Get or set constraints of a qualifyr table}
\usage{
constraints(x, ...)

constraints(x, ...) <- value

\method{constraints}{qf_dataset}(x, table, ...)

\method{constraints}{qf_dataset}(x, table, ...) <- value

\method{constraints}{qf_table}(x, ...)

\method{constraints}{qf_table}(x, dataset = NULL, ...) <- value
}
\arguments{
\item{x}{A qualifyr data set or table to set constraints on}

\item{...}{Additional arguments passed to methods}

\item{value}{A list of constraints (or constraint builder functions)
created with \verb{cstr_*()} functions.}

\item{table}{<\code{\link[=constraints]{tidy-select}}> If \code{x} is a
data set, the table(s) to set constraints on, either as a character string
or bare symbol}

\item{dataset}{If \code{x} is a data frame, the data set used to reference
foreign keys against. Only needed if \code{value} contains a foreign key.}
}
\value{
For \code{constraints}, the list of \verb{<qf_constraint>} objects associated
with the table. For \verb{constraints<-}, an updated version of \code{x} with
constraints set.
}
\description{
Get or set constraints of a qualifyr table
}
