% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_constraint.R, R/qf_dataframe.R,
%   R/qf_dataset.R
\name{constraints}
\alias{constraints}
\alias{constraints<-}
\alias{constraints.qf_dataframe}
\alias{constraints<-.qf_dataframe}
\alias{constraints.qf_dataset}
\alias{constraints<-.qf_dataset}
\title{Get or set constraints of a qualifyr data frame}
\usage{
constraints(x, ...)

constraints(x, ...) <- value

\method{constraints}{qf_dataframe}(x, ...)

\method{constraints}{qf_dataframe}(x, dataset = NULL, ...) <- value

\method{constraints}{qf_dataset}(x, table, ...)

\method{constraints}{qf_dataset}(x, table, ...) <- value
}
\arguments{
\item{x}{A qualifyr data set or data frame to set constraints on}

\item{...}{Additional arguments passed to methods}

\item{value}{A list of constraints (or constraint builder functions)
created with \verb{cstr_*()} functions.}

\item{dataset}{If \code{x} is a data frame, the data set used to reference
foreign keys against. Only needed if \code{value} contains a foreign key.}

\item{table}{If \code{x} is a data set, the name of the table to set constraints
on, either as a character string or bare symbol}
}
\value{
For \code{constraints}, the list of \verb{<qf_constraint>} objects associated
with the table. For \verb{constraints<-}, an updated version of \code{x} with
constraints set.
}
\description{
Get or set constraints of a qualifyr data frame
}
