% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cstr_unique_key.R, R/cstr_primary_key.R,
%   R/cstr_foreign_key.R, R/cstr_not_missing.R, R/qf_constraint.R
\name{cstr_unique_key}
\alias{cstr_unique_key}
\alias{cstr_primary_key}
\alias{cstr_foreign_key}
\alias{cstr_not_missing}
\alias{cstr_}
\title{Constraints for relational data structures}
\usage{
cstr_unique_key(cols)

cstr_primary_key(cols)

cstr_foreign_key(cols, reference)

cstr_not_missing(cols)
}
\arguments{
\item{cols}{<\code{\link[=args_tidy_select]{tidy-select}}> The columns to set
constraints on}

\item{reference}{<\code{\link[=args_tidy_select]{tidy-select}}> For
\code{cstr_foreign_key()}, the table and columns to reference. If the key
columns and reference columns have the same names, you can specify just the
table. Otherwise, specify the columns using \code{table$columns} or
\code{table[columns]} where \code{table} and \code{column}/\code{columns} are each a
tidy-select specification. The special value \code{.self} (as a bare symbol)
refers to the current table.}
}
\value{
A \verb{<qf_constraint_specifier>} object.
}
\description{
Functions to specify constraints for qualifyr tables. The return values
should be stored in a list and passed to \verb{constraints<-}.
}
\details{
The return value, a \verb{<qf_constraint_specifier>} object, is
technically a closure which takes a qualifyr table and returns a
\verb{<qf_constraint>} object. This is a somewhat unfortunate implementation
detail, but it allows using \verb{cstr_*()} functions in the RHS of \code{`constraints<-`()} without needing to include the table as an argument,
making front-end syntex a little clearer.
}
\section{Functions}{
\itemize{
\item \code{cstr_unique_key()}: Requires that each row has a unique set of values for the
key columns.

\item \code{cstr_primary_key()}: Requires that each row has a unique set of values for the
key columns, and is not missing values in the key columns.

\item \code{cstr_foreign_key()}: Requires that set of values in the key matches a set of
values in the reference columns. The reference columns must be a unique key
or a primary key.

\item \code{cstr_not_missing()}: Requires that the specified columns do not contain missing
values.

}}
