% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_constraint.R, R/cstr_unique_key.R,
%   R/cstr_primary_key.R, R/cstr_foreign_key.R
\name{cstr_}
\alias{cstr_}
\alias{cstr_unique_key}
\alias{cstr_primary_key}
\alias{cstr_foreign_key}
\title{Constraints for relational data structures}
\usage{
cstr_unique_key(cols)

cstr_primary_key(cols)

cstr_foreign_key(cols, reference)
}
\arguments{
\item{cols}{<\code{tidy-select}> The columns to set constraints on}

\item{reference}{<\code{tidy-select}> For \code{cstr_foreign_key()}, the table and
columns to reference. If the key columns and reference columns have the
same names, you can specify just the table. Otherwise, specify the columns
using \code{table[columns]} where \code{columns} is a tidy-select specification.}
}
\value{
These functions return a closure of class
\verb{<qf_constraint_specifier>}, which takes a qualifyr table and returns an
object inheriting from \verb{<qf_constraint>}. This unfortunate implementation
detail allows omitting the data set and table when used in the RHS of the
replace-form function \verb{constraints<-}, making front-end syntax a little
nicer.
}
\description{
Functions to specify constraints for qualifyr tables. The
return values should be stored in a list and passed to \verb{constraints<-}.
}
\section{Functions}{
\itemize{
\item \code{cstr_unique_key()}: Requires that each row has a unique set of values for the
key columns.

\item \code{cstr_primary_key()}: Requires that each row has a unique set of values for the
key columns, and is not missing values in the key columns.

\item \code{cstr_foreign_key()}: Requires that set of values in the key matches a set of
values in the reference columns. The reference columns must be a unique key
or a primary key.

}}
