% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_constraint.R
\name{cstr_}
\alias{cstr_}
\alias{cstr_unique_key}
\alias{cstr_primary_key}
\alias{cstr_foreign_key}
\title{Constraints for relational data structures}
\usage{
cstr_unique_key(cols)

cstr_primary_key(cols)

cstr_foreign_key(cols, ref_cols)
}
\arguments{
\item{cols}{The columns to set constraints on}

\item{ref_cols}{For \code{cstr_foreign_key()}, the reference columns for the
foreign key. Specify columns from a different table using \verb{\%from\%}, e.g.,
\code{reference_column \%from\% reference_table}}
}
\value{
These functions return closures of class \verb{<cstr_builder>}, which
take a qualifyr data set and data frame as arguments and return an object
inheriting from \verb{<qf_constraint>}. This implementation detail allows these
functions to be used with the replacement-form function \verb{constraints<-}.
}
\description{
Functions to specify constraints for qualifyr data frames. The
return values should be stored in a list and passed to \verb{constraints<-}.
}
