% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_exception.R
\name{except_where}
\alias{except_where}
\title{Define a constraint exception}
\usage{
except_where(...)
}
\arguments{
\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Conditions under
which to except a row. Each condition, when using the table as a data mask,
should evaluate to a logical vector with length equal to the number of rows
of the table. Rows for which all conditions are \code{TRUE} will be excepted.}
}
\value{
A \verb{<qf_exception>} object
}
\description{
Define a constraint exception
}
