% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_table.R, R/cstr_unique_key.R,
%   R/cstr_primary_key.R, R/cstr_foreign_key.R
\name{pick_constraint}
\alias{pick_constraint}
\alias{unique_key}
\alias{primary_key}
\alias{foreign_key}
\alias{unique_key<-}
\alias{primary_key<-}
\alias{foreign_key<-}
\title{Get or replace a specific constraint}
\usage{
unique_key(table, cols = NULL)

primary_key(table, cols = NULL)

foreign_key(table, cols = NULL)

unique_key(table, cols = NULL) <- value

primary_key(table, cols = NULL) <- value

foreign_key(table, cols = NULL) <- value
}
\arguments{
\item{table}{The table to search for the constraint}

\item{cols}{\if{html}{\out{<tidy-select>}} The columns the constraint applies to. If \code{cols} is
\code{NULL} or otherwise identifies zero columns, returns the constraint with
the specified type if there is exactly one, and throws an error otherwise.}

\item{value}{The constraint object to replace the specified constraint with}
}
\value{
For the selection forms, the constraint object. For the replacement
forms, the updated table object.
}
\description{
These functions let you identify a specific constraint on a
table for the purpose of inspecting, modifying, or replacing it.
}
