% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_dataset.R
\name{qf_dataset}
\alias{qf_dataset}
\title{Create a new qualifyr data set from data frames}
\usage{
qf_dataset(...)
}
\arguments{
\item{...}{Data frames to be included in the data set. If names are provided,
they will be used as the table names; otherwise, the deparsed expressions
will be used, which may or may not be a sensible default.}
}
\value{
A new \verb{<qf_dataset>} object containing the tables provided in \code{...}
(after being converted to \verb{<qf_table>} objects by \code{as_qf_table()}).
}
\description{
Create a new qualifyr data set from data frames
}
