% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/qf_constraint.R, R/qf_dataset.R,
%   R/qf_table.R
\name{type-predicates}
\alias{type-predicates}
\alias{is_qf_constraint}
\alias{is_qf_dataset}
\alias{is_qf_table}
\title{Test for qualifyr objects}
\usage{
is_qf_constraint(x)

is_qf_dataset(x)

is_qf_table(x)
}
\arguments{
\item{x}{The object to test the type of}
}
\value{
\code{TRUE} if \code{x} inherits from the specified class; \code{FALSE} otherwise.
}
\description{
Test for qualifyr objects
}
